# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  country_dto.CountryResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        additionalProperties:
          type: string
        type: object
      parent_id:
        type: integer
      updated_at:
        type: string
    type: object
  country_dto.CreateCountryRequest:
    properties:
      is_active:
        type: boolean
      name:
        additionalProperties:
          type: string
        type: object
      parent_id:
        minimum: 1
        type: integer
    required:
    - name
    type: object
  country_dto.UpdateCountryRequest:
    properties:
      is_active:
        type: boolean
      name:
        additionalProperties:
          type: string
        type: object
      parent_id:
        minimum: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Admin paneli uchun CRUD endpointlari (cursor pagination, filter, sort)
  title: Usta Backend Admin API
  version: "1.0"
paths:
...
  /countries:
    get:
      parameters:
      - description: Keyset cursor
        in: query
        name: cursor
        type: string
      - description: Default 10, max 100
        in: query
        name: limit
        type: integer
      - description: id|name|parent_id|is_active|created_at|updated_at
        in: query
        name: sort_by
        type: string
      - description: asc|desc
        in: query
        name: sort_order
        type: string
      - description: name JSONB ichida ILIKE
        in: query
        name: name
        type: string
      - description: Aniq parent_id
        in: query
        name: parent_id
        type: integer
      - description: true => parent_id NOT NULL, false => root
        in: query
        name: has_parent
        type: boolean
      - description: Holat
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Countries ro'yxati (admin, cursor pagination)
      tags:
      - Countries
    post:
      consumes:
      - application/json
      description: 'name maydoni JSONB: default + har bir faol til kalitlari majburiy'
      parameters:
      - description: Country ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/country_dto.CreateCountryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/country_dto.CountryResponse'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Davlat/region yaratish (admin, ko'p tilli)
      tags:
      - Countries
  /countries/{id}:
    delete:
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Countryni o'chirish (soft delete, admin)
      tags:
      - Countries
    get:
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/country_dto.CountryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Countryni ID bo'yicha olish (admin)
      tags:
      - Countries
    put:
      consumes:
      - application/json
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yangilanadigan maydonlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/country_dto.UpdateCountryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/country_dto.CountryResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Countryni yangilash (admin)
      tags:
      - Countries
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"